refactor: move repo params to types/repo_params.rs and split CI test clients
Move all *Params structs from repository.rs into a new
types/repo_params.rs module for better organization.

Rework integration test client setup for CI:
- prod_api() uses HF_PROD_TOKEN against huggingface.co for read-only tests
- api() uses HF_CI_TOKEN against hub-ci.huggingface.co for write tests
- Local dev uses HF_TOKEN with default/env endpoint via default_api()
- All client creation goes through build_client() with explicit token+endpoint
