+++ huggingface_hub/src/types/mod.rs
@@ -2,6 +2,7 @@ pub mod cache;
 pub mod commit;
 pub mod params;
 pub mod repo;
+pub mod repo_params;
 pub mod user;
 
 #[cfg(feature = "spaces")]
@@ -10,6 +11,7 @@ pub mod spaces;
 pub use commit::*;
 pub use params::*;
 pub use repo::*;
+pub use repo_params::*;
 #[cfg(feature = "spaces")]
 pub use spaces::*;
 pub use user::*;
